+++ etils/epy/__init__.py
@@ -14,6 +14,8 @@
 
 """Python utils public API."""
 
+# pylint: disable=g-importing-member
+
 import sys
 
 from etils.epy.backports import cached_property
@@ -22,6 +24,7 @@ from etils.epy.env_utils import is_notebook
 from etils.epy.itertools import groupby
 from etils.epy.itertools import splitby
 from etils.epy.itertools import zip_dict
+from etils.epy.lazy_imports_utils import lazy_imports
 from etils.epy.py_utils import frozen
 from etils.epy.py_utils import issubclass_ as issubclass  # pylint: disable=redefined-builtin
 from etils.epy.py_utils import StrEnum